requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
cffi==1.16.0
cryptography==42.0.7
diff-match-patch==20230430
Django==4.2.11
django-ckeditor==6.7.1
django-import-export==4.0.0
django-js-asset==2.2.0
djangorestframework==3.15.1
pillow==10.3.0
pycparser==2.22
PyMySQL==1.1.0
sqlparse==0.5.0
tablib==3.5.0
typing-extensions==4.11.0
